maths part works well, still issues with texture

+++ glimac/include/glimac/RadialBasisFunctions.hpp
@@ -52,5 +52,5 @@ std::vector<glm::vec2> get_control_points_RBF(const std::string &filename);
 Eigen::VectorXd get_omega_variables(std::vector<glm::vec2> control_points, Eigen::VectorXd u_vect, 
                                     phi_functors phi_function, const unsigned int type_function);
                                     
-std::vector<unsigned int> interpolate(std::vector<glm::vec2> control_points, Eigen::VectorXd omegas, 
+std::vector<float> interpolate(std::vector<glm::vec2> control_points, Eigen::VectorXd omegas, 
                                       std::vector<glm::vec2> all_positions, phi_functors phi_function, const unsigned int type_function);
